refactor: SSE 알림 단순화 및 중복 검사 제거
- 알림 상세 데이터는 목록 조회에서만 제공
- 새 알림 생성 시 뱃지 증가 이벤트만 실시간 전송
- 알림 중복 검사 제거 → 탈퇴 후 재가입 등 정상 케이스 허용
