+++ src/notifications/types/notification-response.type.ts
