+++ src/notifications/types/notification-signal.type.ts
@@ -1,0 +1,3 @@
+export interface NotificationSignal {
+  type: 'NEW_NOTIFICATION';
+}
